121: Refactor, error handling
