+++ backend/src/reddit_fetcher/mod.rs
@@ -1,0 +1,6 @@
+mod fetcher_error;
+pub mod nlp_request;
+
+pub mod fetcher;
+pub(crate) mod model;
+pub mod reddit;
