+++ backend/src/reddit_fetcher/model/mod.rs
@@ -1,0 +1,4 @@
+pub mod post_comments;
+pub mod posts;
+pub mod reddit_data;
+pub mod user_posts;
