+++ backend/src/reddit_fetcher/reddit/request/tests.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use crate::reddit::request::{
+    use crate::reddit_fetcher::reddit::request::{
         params::{FeedSorting, FeedSortingTime},
         PostCommentsRequest, RedditResource, SubredditInfoRequest, SubredditPostsRequest,
         UserInfoRequest, UserPostsRequest,
@@ -17,7 +17,7 @@ mod tests {
             sorting: FeedSorting::New,
             after: None,
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(url, "https://oauth.reddit.com/r/Polska/new.json");
         assert_eq!(query, vec![("limit", "100".to_string())]);
     }
@@ -27,7 +27,7 @@ mod tests {
         let req = SubredditInfoRequest {
             subreddit: "Polska".to_string(),
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(url, "https://oauth.reddit.com/r/Polska/about.json");
         assert_eq!(query, vec![]);
     }
@@ -39,7 +39,7 @@ mod tests {
             sorting: FeedSorting::Top(FeedSortingTime::All),
             after: None,
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(url, "https://oauth.reddit.com/user/spez.json");
         assert_eq!(
             query,
@@ -56,7 +56,7 @@ mod tests {
         let req = UserInfoRequest {
             username: "spez".to_string(),
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(url, "https://oauth.reddit.com/user/spez/about.json");
         assert_eq!(query, vec![]);
     }
@@ -69,7 +69,7 @@ mod tests {
             sorting: FeedSorting::Controversial(FeedSortingTime::Day),
             after: None,
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(
             url,
             "https://oauth.reddit.com/r/Polska/comments/abc123.json"
@@ -92,7 +92,7 @@ mod tests {
             sorting: FeedSorting::default(),
             after: None,
         };
-        let (url, query) = req.into_request_parts();
+        let (url, query) = req.to_request_parts();
         assert_eq!(
             url,
             "https://oauth.reddit.com/r/Polska/comments/abc123.json"
